+++ src/main/java/Dao/DAODelete.java
@@ -1,0 +1,5 @@
+package Dao;
+
+public interface DAODelete<T> {
+    void delete(T t);
+}
