+++ src/main/java/Dao/DAOInterface.java
@@ -6,6 +6,6 @@ public interface DAOInterface<T> {
     List<T> getAll();
     void create(T t);
     void update(T t, String[] params);
-    void delete(T t);
+
 
 }
