+++ src/main/java/Dao/OrderDao.java
@@ -8,7 +8,7 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class OrderDao extends AbstractDao<Order> {
+public class OrderDao extends AbstractDao<Order> implements DAODelete<Order>{
 
 
     @Override
@@ -32,7 +32,7 @@ public class OrderDao extends AbstractDao<Order> {
     public List<Order> getAll() {
         List<Order> orders = new ArrayList<>();
         String SQL_SELECT_ALL = "Select* from orders";
-        try(PreparedStatement ps=this.executePrepareStatement(SQL_SELECT_ALL)) {
+        try (PreparedStatement ps=this.executePrepareStatement(SQL_SELECT_ALL)) {
             ResultSet rs=ps.executeQuery();
             while(rs.next()){
                 orders.add(new Order(rs.getLong("order_id"),rs.getLong("customer_id"),
@@ -43,13 +43,13 @@ public class OrderDao extends AbstractDao<Order> {
         } catch(SQLException e){
             e.printStackTrace();
         }
-        return orders;
+        return null;
     }
 
     @Override
     public void create(Order order) {
         String SQL_INSERT = "Insert into orders (customer_id, order_date, total_amount) values(?,?,?)";
-        try(PreparedStatement ps= this.executePrepareStatement(SQL_INSERT)) {
+        try (PreparedStatement ps= this.executePrepareStatement(SQL_INSERT)) {
             ps.setLong(1, order.getCustomerId());
             ps.setDate(2, order.getOrderDate());
             ps.setLong(3, order.getTotalAmount());
@@ -62,7 +62,7 @@ public class OrderDao extends AbstractDao<Order> {
     @Override
     public void update(Order order, String[] params) {
         String SQL_UPDATE="Update order set order_id=?,order_date=?,customer_id=?, total_amount=? where order_id=?";
-        try(PreparedStatement ps=this.executePrepareStatement(SQL_UPDATE)){
+        try (PreparedStatement ps=this.executePrepareStatement(SQL_UPDATE)){
             ps.setString(1, params[0]);
             ps.setString(2, params[1]);
             ps.setString(3, params[2]);
@@ -77,13 +77,11 @@ public class OrderDao extends AbstractDao<Order> {
     @Override
     public void delete(Order order) {
         String SQL_DELETE="Delete from order where order_id=?";
-        try(PreparedStatement ps=this.executePrepareStatement(SQL_DELETE)){
+        try (PreparedStatement ps=this.executePrepareStatement(SQL_DELETE)){
             ps.setLong(1, order.getOrderId());
             ps.executeUpdate();
         } catch(SQLException e){
             e.printStackTrace();
         }
     }
-
-
 }
