+++ src/main/java/Services/CRUService.java
@@ -1,0 +1,11 @@
+package Services;
+
+import java.util.List;
+
+public interface CRUService<T> {
+    T getById(int id);
+    List<T> getAll();
+    void create(T t);
+    void update(T t, String[] params);
+
+}
