+++ src/main/java/Services/DeleteService.java
@@ -1,0 +1,5 @@
+package Services;
+
+public interface DeleteService<T> {
+    void delete(T t);
+}
