+++ src/main/java/Services/ServiceImpl/ReviewCRUImpl.java
@@ -1,0 +1,45 @@
+package Services.ServiceImpl;
+
+import Dao.ReviewDao;
+import Entities.Review;
+import Services.CRUService;
+import Services.DeleteService;
+
+import java.util.List;
+
+public class ReviewCRUImpl implements CRUService<Review>, DeleteService<Review> {
+
+    ReviewDao reviewDao = null;
+
+    public ReviewCRUImpl(ReviewDao reviewDao) {
+        this.reviewDao = reviewDao;
+    }
+
+    @Override
+    public Review getById(int id) {
+        return reviewDao.getById(id);
+    }
+
+    @Override
+    public List<Review> getAll() {
+        return reviewDao.getAll();
+    }
+
+    @Override
+    public void create(Review review) {
+        reviewDao.create(review);
+
+    }
+
+    @Override
+    public void update(Review review, String[] params) {
+        reviewDao.update(review, params);
+
+    }
+
+    @Override
+    public void delete(Review review) {
+        reviewDao.delete(review);
+
+    }
+}
